Keep the isolation tests in the existing Gitea suite
The multi-tenant tests move into VCSGiteaConsoleClientTest and the OAuth2
dance takes its project, headers and Gitea account as parameters in place, so
the suite gains the coverage without relocating helpers or adding a class.
